fix: created get_bmi seperate file
